all node folders and main cpp h files
