+++ march_ws/build.sh
@@ -12,4 +12,4 @@ catkin_lint src/*/ --ignore missing_directory --ignore literal_project_name --ig
 source devel/setup.bash || exit 1
 catkin_make run_tests && catkin_test_results || exit 1
 
-
+rosrun roslaunch roslaunch-check src/*/launch || exit 1
