+++ march_ws/src/march_main/src/control_node/main.h
@@ -1,0 +1,9 @@
+// Copyright 2018 Project March.
+#ifndef SRC_CONTROL_NODE_MAIN_H_
+#define SRC_CONTROL_NODE_MAIN_H_
+
+class main
+{
+};
+
+#endif  // SRC_CONTROL_NODE_MAIN_H_
