--- march_ws/src/march_main/src/data_logger.h
@@ -1,10 +1,0 @@
-// Copyright 2018 Project March.
-
-#ifndef MARCH_MAIN_DATA_LOGGER_H
-#define MARCH_MAIN_DATA_LOGGER_H
-
-class data_logger
-{
-};
-
-#endif  // MARCH_MAIN_DATA_LOGGER_H
