+++ march_ws/src/march_main/src/datalogger_node/main.h
@@ -1,0 +1,10 @@
+// Copyright 2018 Project March.
+
+#ifndef SRC_DATALOGGER_NODE_MAIN_H_
+#define SRC_DATALOGGER_NODE_MAIN_H_
+
+class data_logger
+{
+};
+
+#endif  // SRC_DATALOGGER_NODE_MAIN_H_
