--- march_ws/src/march_main/src/developer_input_node.h
@@ -1,10 +1,0 @@
-// Copyright 2018 Project March.
-
-#ifndef MARCH_MAIN_DEVELOPER_INPUT_NODE_H
-#define MARCH_MAIN_DEVELOPER_INPUT_NODE_H
-
-class developer_input_node
-{
-};
-
-#endif  // MARCH_MAIN_DEVELOPER_INPUT_NODE_H
