+++ march_ws/src/march_main/src/developer_input_node/main.cpp
@@ -1,9 +1,9 @@
 // Copyright 2018 Project March.
 
-#include "developer_input_node.h"
+#include "main.h"
 #include "std_msgs/Bool.h"
 #include "ros/ros.h"
-#include "public/enum/gait_enum.h"
+#include "../public/enum/gait_enum.h"
 #include <march_custom_msgs/Gait.h>
 
 int main(int argc, char** argv)
