+++ march_ws/src/march_main/src/developer_input_node/main.h
@@ -1,0 +1,10 @@
+// Copyright 2018 Project March.
+
+#ifndef SRC_DEVELOPER_INPUT_NODE_MAIN_H
+#define SRC_DEVELOPER_INPUT_NODE_MAIN_H
+
+class developer_input_node
+{
+};
+
+#endif  // SRC_DEVELOPER_INPUT_NODE_MAIN_H
