--- march_ws/src/march_main/src/gait_controller_node.h
@@ -1,10 +1,0 @@
-// Copyright 2018 Project March.
-
-#ifndef MARCH_MAIN_GAIT_CONTROLLER_NODE_H
-#define MARCH_MAIN_GAIT_CONTROLLER_NODE_H
-
-class gait_controller_node
-{
-};
-
-#endif  // MARCH_MAIN_GAIT_CONTROLLER_NODE_H
