+++ march_ws/src/march_main/src/gait_node/main.h
@@ -1,0 +1,10 @@
+// Copyright 2018 Project March.
+
+#ifndef SRC_GAIT_NODE_MAIN_H
+#define SRC_GAIT_NODE_MAIN_H
+
+class gait_controller_node
+{
+};
+
+#endif  // SRC_GAIT_NODE_MAIN_H
