--- march_ws/src/march_main/src/march_main_node/Cube.cpp
@@ -1,12 +1,0 @@
-// Copyright 2018 Project March.
-
-#include "Cube.h"
-
-TestCube::TestCube(float volume) : volume(volume)
-{
-}
-
-float TestCube::getVolume() const
-{
-  return volume;
-}
