--- march_ws/src/march_main/src/march_main_node/main.h
@@ -1,10 +1,0 @@
-// Copyright 2018 Project March.
-
-#ifndef MARCH_WS_SRC_MARCH_MAIN_SRC_MARCH_MAIN_NODE_MAIN_H_
-#define MARCH_WS_SRC_MARCH_MAIN_SRC_MARCH_MAIN_NODE_MAIN_H_
-
-class march_main_node
-{
-public:
-};
-#endif  // MARCH_WS_SRC_MARCH_MAIN_SRC_MARCH_MAIN_NODE_MAIN_H_
