+++ march_ws/src/march_main/src/master_node/main.h
@@ -1,0 +1,6 @@
+// Copyright 2018 Project March.
+
+#ifndef SRC_MASTER_NODE_MAIN_H
+#define SRC_MASTER_NODE_MAIN_H
+
+#endif  // SRC_MASTER_NODE_MAIN_H
