--- march_ws/src/march_main/src/master_node/master_node.h
@@ -1,6 +1,0 @@
-// Copyright 2018 Project March.
-
-#ifndef MARCH_MAIN_MASTER_NODE_H
-#define MARCH_MAIN_MASTER_NODE_H
-
-#endif  // MARCH_MAIN_MASTER_NODE_H
