+++ march_ws/src/march_main/src/sensor_node/main.cpp
@@ -1,0 +1,10 @@
+// Copyright 2018 Project March.
+#include "ros/ros.h"
+#include "main.h"
+
+int main(int argc, char** argv)
+{
+  ros::init(argc, argv, "sensor_node");
+  ros::NodeHandle n;
+  return 0;
+}
